Prevent re-mounting of the child component
When specifying a url prop, I noticed that my child component was mounted, unmounted and then mounted again. Since I'm doing network stuff in my componentDidMount, I can't really accept that. I'm not sure why the onloadHandler needs to be called again, but it seems logical to just use the existing container if that is the case. componentDidMount will only be called once. I guess there's a case where someone sets containerId via props, so you need to unmount the component with that containId and then remount it?